+++ tests/ARCTokenization.Tests/TestObjects.fs
@@ -2,24 +2,8 @@
 
 open ARCTokenization
 open FsSpreadsheet
-
-module Tokenization =
-    
-    module InvestigationMetadata =
-        
-        ()
-
-module Integration =
-    
-    module AssayAnnotationTable =
-    
-        module Correct =
-        
-            let ``assay with only source and sample column`` =
-                Assay.parseAnnotationTablesFromFile "Fixtures/correct/assay_with_only_source_and_sample_column.xlsx"
-
-            let ``assay with single characteristics`` = 
-                Assay.parseAnnotationTablesFromFile "Fixtures/correct/assay_with_single_characteristics.xlsx"
+open ARCtrl
+open ARCtrl.ISA
 
 module MockAPI =
     
@@ -43,3 +27,52 @@ module MockAPI =
         let empty = 
             ARCMock.AssayMetadataTokens()
             |> List.concat // use flat list
+
+    module ProcessGraphTokens =
+        
+        let inputColumn = 
+            ARCMock.ProcessGraphColumn(
+                header = CompositeHeader.Input IOType.Source,
+                cells = [
+                    CompositeCell.FreeText "Source_1"
+                    CompositeCell.FreeText "Source_1"
+                ]
+            )
+
+        let characteristicsColumn =
+            ARCMock.ProcessGraphColumn(
+                header = CompositeHeader.Characteristic (OntologyAnnotation.create(TermAccessionNumber = "OBI:0100026", Name = AnnotationValue.Text "organism", TermSourceREF = "OBI")),
+                cells = [
+                    CompositeCell.Term (OntologyAnnotation.create(TermAccessionNumber = "http://purl.obolibrary.org/obo/NCBITaxon_3702", Name = AnnotationValue.Text "Arabidopsis thaliana", TermSourceREF = "NCBITaxon"))
+                    CompositeCell.Term (OntologyAnnotation.create(TermAccessionNumber = "http://purl.obolibrary.org/obo/NCBITaxon_3702", Name = AnnotationValue.Text "Arabidopsis thaliana", TermSourceREF = "NCBITaxon"))
+                ]
+            )
+                    
+        let outputColumn = 
+            ARCMock.ProcessGraphColumn(
+                header = CompositeHeader.Output IOType.Sample,
+                cells = [
+                    CompositeCell.FreeText "Sample_1"
+                    CompositeCell.FreeText "Sample_2"
+                ]
+            )
+
+        let simpleStudy = 
+            ARCMock.ProcessGraph(
+                [
+                    CompositeHeader.Input IOType.Source, [
+                        CompositeCell.FreeText "Source_1"
+                        CompositeCell.FreeText "Source_1"
+                    ]
+
+                    CompositeHeader.Characteristic (OntologyAnnotation.create(TermAccessionNumber = "OBI:0100026", Name = AnnotationValue.Text "organism", TermSourceREF = "OBI")), [
+                        CompositeCell.Term (OntologyAnnotation.create(TermAccessionNumber = "http://purl.obolibrary.org/obo/NCBITaxon_3702", Name = AnnotationValue.Text "Arabidopsis thaliana", TermSourceREF = "NCBITaxon"))
+                        CompositeCell.Term (OntologyAnnotation.create(TermAccessionNumber = "http://purl.obolibrary.org/obo/NCBITaxon_3702", Name = AnnotationValue.Text "Arabidopsis thaliana", TermSourceREF = "NCBITaxon"))
+                    ]
+
+                    CompositeHeader.Output IOType.Sample, [
+                        CompositeCell.FreeText "Sample_1"
+                        CompositeCell.FreeText "Sample_2"
+                    ]
+                ]
+            )