+++ tests/ARCTokenization.Tests/TokenizationTests/ProcessGraphTokenization.fs
@@ -1,0 +1,6 @@
+﻿namespace TokenizationTests.ProcessGraph
+
+open ControlledVocabulary
+open ARCTokenization
+open Xunit
+